refactor: consolidate combo data scripts
