+++ package.json
@@ -9,9 +9,8 @@
     "prebuild": "prettier --write .",
     "build": "nuxt build",
     "start": "nuxt start",
-    "copy-google-sheets": "npx ts-node scripts/copy-google-sheets",
     "download-external-card-data": "npx ts-node scripts/download-external-card-data",
-    "download-data": "npm run copy-google-sheets && npm run download-external-card-data",
+    "download-data": "npx ts-node scripts/download-data",
     "pregenerate": "npm run download-data",
     "generate": "nuxt generate --fail-on-error",
     "lint:js": "eslint --ext .ts,.vue .",
