+++ scripts/download-data/get-edhrec-prices.ts
