+++ scripts/download-data/index.ts
@@ -2,6 +2,8 @@ import fs from "fs";
 import log from "../shared/log";
 import getScryfallData from "./get-scryfall";
 import getEDHRecPrices from "./get-edhrec-prices";
+import getGoogleSheetsComboData from "./get-google-sheets-data";
+import { collectCardNames, collectResults } from "./collect-autocomplete";
 
 type CardData = {
   name: string;
@@ -15,9 +17,23 @@ type CardData = {
   };
 };
 
-Promise.all([getScryfallData(), getEDHRecPrices()]).then((responses) => {
+Promise.all([
+  getGoogleSheetsComboData(),
+  getScryfallData(),
+  getEDHRecPrices(),
+]).then((responses) => {
   const data = [] as CardData[];
-  const [scryfallData, edhrecData] = responses;
+  const [compressedData, scryfallData, edhrecData] = responses;
+  const cardNames = collectCardNames(compressedData);
+  const results = collectResults(compressedData);
+
+  log("Writing /autocomplete-data/cards.json");
+  fs.writeFileSync("./autocomplete-data/cards.json", JSON.stringify(cardNames));
+  log("/autocomplete-data/cards.json written", "green");
+
+  log("Writing /autocomplete-data/results.json");
+  fs.writeFileSync("./autocomplete-data/results.json", JSON.stringify(results));
+  log("/autocomplete-data/results.json written", "green");
 
   Object.keys(scryfallData).forEach((cardName) => {
     const priceData = edhrecData[cardName] || {
@@ -44,4 +60,11 @@ Promise.all([getScryfallData(), getEDHRecPrices()]).then((responses) => {
       log(`${fileName} written`, "green");
     });
   });
+
+  log("Writing /static/api/combo-data.json");
+  fs.writeFileSync(
+    "./static/api/combo-data.json",
+    JSON.stringify(compressedData)
+  );
+  log("/static/api/combo-data.json written", "green");
 });
